+++ unittest/src/fake/syscall_driver.rs
@@ -25,7 +25,7 @@ pub trait SyscallDriver: 'static {
 
     /// Process a Command system call. Fake drivers should use the methods in
     /// `libtock_unittest::command_return` to construct the return value.
-    fn command(&self, command_id: u32, argument0: u32, argument1: u32) -> CommandReturn;
+    fn command(&self, command_id: u32, argument0: usize, argument1: usize) -> CommandReturn;
 
     // -------------------------------------------------------------------------
     // Allow
